system/vold: Fix the issue of not able to mount 32 GB SD card after formatting it from UI.
Cluster for sector option is set to 8 while formatting SD card.
System is taking more memory for checking FAT on 32 GB SD card
because of more number of clusters. System is killing the applications
due to low memory. Android framework is getting rebooted when the Launcher
application is killed and SD card is not able mount.

Fix the issue by suppressing the sectors per cluster options while calling
the format utility. Format utility calculate sectors per cluster based on
the SD card size and number of clusters are optimized.

Change-Id: I5e4670ad13018014a7a2f0027f44bcaa899b70ba
CRs-fixed: 248144